Feature/tratamento erros

+++ src/main/java/com/senai/projetonotas/service/impl/ProfessorServiceImpl.java
@@ -1,6 +1,9 @@
 package com.senai.projetonotas.service.impl;
 
 import com.senai.projetonotas.entity.ProfessorEntity;
+import com.senai.projetonotas.exception.customException.CampoObrigatorioException;
+import com.senai.projetonotas.exception.customException.NotFoundException;
+import com.senai.projetonotas.repository.DisciplinaRepository;
 import com.senai.projetonotas.repository.ProfessorRepository;
 import com.senai.projetonotas.service.ProfessorService;
 import lombok.RequiredArgsConstructor;
@@ -14,9 +17,14 @@ public class ProfessorServiceImpl implements ProfessorService {
 
 
     private final ProfessorRepository professorRepository;
+    private final DisciplinaRepository disciplinaRepository;
 
     @Override
     public ProfessorEntity create(ProfessorEntity professor) {
+
+        if (professor.getNome() == null) {
+            throw new CampoObrigatorioException("O campo 'nome' é obrigatório para criar um professor");
+        }
         return professorRepository.save(professor);
     }
 
@@ -27,19 +35,26 @@ public class ProfessorServiceImpl implements ProfessorService {
 
     @Override
     public ProfessorEntity getEntity(Long id) {
-        return professorRepository.findById(id).orElseThrow(() -> new RuntimeException("Professor não encontrado"));
+        return professorRepository.findById(id).orElseThrow(() -> new NotFoundException("Não econtrado professor de id: " +id));
     }
 
     @Override
     public ProfessorEntity update(Long id, ProfessorEntity professor) {
         getEntity(id);
         professor.setProfessorId(id);
-        return professorRepository.save(professor);
+        return create(professor);
     }
 
     @Override
     public void delete(Long id) {
         getEntity(id);
+
+        long disciplinasProfessor = disciplinaRepository.countByProfessorProfessorId(id);
+
+        if (disciplinasProfessor > 0) {
+            throw new IllegalStateException("Não é possível excluir o professor porque existem disciplinas associadas a ele.");
+        }
+
         professorRepository.deleteById(id);
 
     }
